+++ src/constants/fields.ts
@@ -1,0 +1,27 @@
+/**
+ * Fields.
+ * @module constants/fields
+ */
+
+export default {
+  label_choice_field: 'Choice',
+  label_date_field: 'Date',
+  label_datetime_field: 'Date/Time',
+  Email: 'Email',
+  label_email: 'Email',
+  'File Upload': 'File Upload',
+  label_float_field: 'Floating-point number',
+  label_full_name: 'Full Name',
+  Image: 'Image',
+  label_integer_field: 'Integer',
+  JSONField: 'JSONField',
+  label_multi_choice_field: 'Multiple Choice',
+  label_password_field: 'Password',
+  'Relation Choice': 'Relation Choice',
+  'Relation List': 'Relation List',
+  'Rich Text': 'Rich Text',
+  label_text_field: 'Text',
+  label_textline_field: 'Text line (String)',
+  URL: 'URL',
+  label_boolean_field: 'Yes/No',
+};