+++ src/vocabularies/index.ts
@@ -8,6 +8,7 @@ import { actions } from './actions/actions';
 import { availableLanguages } from './available-languages/available-languages';
 import { behaviors } from './behaviors/behaviors';
 import { captchaProviders } from './captcha-providers/captcha-providers';
+import { fields } from './fields/fields';
 import { groups } from './groups/groups';
 import { imageScales } from './image-scales/image-scales';
 import { permissions } from './permissions/permissions';
@@ -29,6 +30,7 @@ export const vocabularies = {
   availableLanguages,
   behaviors,
   captchaProviders,
+  Fields: fields,
   groups,
   imageScales,
   permissions,