+++ CHANGELOG.md
@@ -117,6 +117,7 @@
 - Add context navigation endpoint @robgietema
 - Add depth option to navigation endpoint @robgietema
 - Add commenting @robgietema
+- Add fields vocabulary @robgietema
 
 ### Bugfix